+++ sample-apps/react-native/ringing-tutorial/app/(app)/_layout.tsx
@@ -1,8 +1,7 @@
 import { Text } from 'react-native';
-import { Redirect, router, Stack } from 'expo-router';
+import { Redirect, Stack } from 'expo-router';
 import { useAuthentication } from '../../contexts/authentication-provider';
-import { useCalls } from '@stream-io/video-react-native-sdk';
-import { useEffect } from 'react';
+import { GestureHandlerRootView } from 'react-native-gesture-handler';
 
 export default function AppLayout() {
   const { userWithToken, isLoading } = useAuthentication();
@@ -15,18 +14,9 @@ export default function AppLayout() {
     return <Redirect href="/login" />;
   }
 
-  return <RingingWrappedStack />;
+  return (
+    <GestureHandlerRootView>
+      <Stack screenOptions={{ headerShown: false }} />
+    </GestureHandlerRootView>
+  );
 }
-
-const RingingWrappedStack = () => {
-  const calls = useCalls().filter((c) => c.ringing);
-
-  useEffect(() => {
-    // whenever there is a ringing call, redirect to the ringing screen
-    if (calls[0]) {
-      router.replace('/(app)/ringing');
-    }
-  }, [calls]);
-
-  return <Stack screenOptions={{ headerShown: false }} />;
-};